Rename `overview-glue-records` to `domain-glue-records`
